Release 6.2.0

+++ tests/modules/Administration/Bug36978Test.php
@@ -1,4 +1,38 @@
 <?php
+/*********************************************************************************
+ * SugarCRM Community Edition is a customer relationship management program developed by
+ * SugarCRM, Inc. Copyright (C) 2004-2011 SugarCRM Inc.
+ * 
+ * This program is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Affero General Public License version 3 as published by the
+ * Free Software Foundation with the addition of the following permission added
+ * to Section 15 as permitted in Section 7(a): FOR ANY PART OF THE COVERED WORK
+ * IN WHICH THE COPYRIGHT IS OWNED BY SUGARCRM, SUGARCRM DISCLAIMS THE WARRANTY
+ * OF NON INFRINGEMENT OF THIRD PARTY RIGHTS.
+ * 
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more
+ * details.
+ * 
+ * You should have received a copy of the GNU Affero General Public License along with
+ * this program; if not, see http://www.gnu.org/licenses or write to the Free
+ * Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
+ * 02110-1301 USA.
+ * 
+ * You can contact SugarCRM, Inc. headquarters at 10050 North Wolfe Road,
+ * SW2-130, Cupertino, CA 95014, USA. or at email address contact@sugarcrm.com.
+ * 
+ * The interactive user interfaces in modified source and object code versions
+ * of this program must display Appropriate Legal Notices, as required under
+ * Section 5 of the GNU Affero General Public License version 3.
+ * 
+ * In accordance with Section 7(b) of the GNU Affero General Public License version 3,
+ * these Appropriate Legal Notices must retain the display of the "Powered by
+ * SugarCRM" logo. If the display of the logo is not reasonably feasible for
+ * technical reasons, the Appropriate Legal Notices must display the words
+ * "Powered by SugarCRM".
+ ********************************************************************************/
 
 class Bug36978Test extends Sugar_PHPUnit_Framework_TestCase {
 
@@ -6,13 +40,10 @@ var $rel_guid;
 var $has_custom_table_dictionary;	
 var $moduleList;
 
-function setUp() {
-	
-    if(true) {
-       $this->markTestSkipped("Skipping unless otherwise specified");
-       return;
-    }	
-	
+function setUp() 
+{
+    $this->markTestSkipped("Skipping unless otherwise specified");
+    
     $admin = new User();
     $GLOBALS['current_user'] = $admin->retrieve('1');	
 	
@@ -234,17 +265,33 @@ function tearDown() {
        unlink('modules/abc_Test/metadata/studio.php'); 
     }    
     
-	rmdir_recursive('custom/Extension/modules/abc_Test/Ext/Vardefs');
-	rmdir_recursive('custom/Extension/modules/abc_Test/Ext/Layoutdefs');
-	rmdir_recursive('custom/Extension/modules/abc_Test/Ext');
-	rmdir_recursive('custom/Extension/modules/abc_Test');
-	rmdir_recursive('modules/abc_Test/metadata');
-	rmdir_recursive('modules/abc_Test');
+    if(is_dir('custom/Extension/modules/abc_Test/Ext/Vardefs')) {
+        rmdir_recursive('custom/Extension/modules/abc_Test/Ext/Vardefs');
+    }
+    if(is_dir('custom/Extension/modules/abc_Test/Ext/Layoutdefs')) {
+        rmdir_recursive('custom/Extension/modules/abc_Test/Ext/Layoutdefs');
+    }
+    if(is_dir('custom/Extension/modules/abc_Test/Ext')) {
+        rmdir_recursive('custom/Extension/modules/abc_Test/Ext');
+    }
+    if(is_dir('custom/Extension/modules/abc_Test')) {
+        rmdir_recursive('custom/Extension/modules/abc_Test');
+    }
+    if(is_dir('modules/abc_Test/metadata')) {
+        rmdir_recursive('modules/abc_Test/metadata');
+    }
+    if(is_dir('modules/abc_Test')) {
+        rmdir_recursive('modules/abc_Test');
+    }
 	
-	$sql = "DELETE FROM relationships WHERE id = '{$this->rel_guid}'";
-	$GLOBALS['db']->query($sql);
+    if ( !empty($this->rel_guid) ) {
+        $sql = "DELETE FROM relationships WHERE id = '{$this->rel_guid}'";
+        $GLOBALS['db']->query($sql);
+	}
 	
-	$GLOBALS['moduleList'] = $this->moduleList;
+	if ( !empty($this->moduleList) ) {
+        $GLOBALS['moduleList'] = $this->moduleList;
+    }
 }
